+++ rpc/backend/feemarket_query_client_test.go
@@ -1,0 +1,21 @@
+package backend
+
+import (
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	"github.com/evmos/ethermint/rpc/backend/mocks"
+	rpc "github.com/evmos/ethermint/rpc/types"
+	feemarkettypes "github.com/evmos/ethermint/x/feemarket/types"
+)
+
+var _ feemarkettypes.QueryClient = &mocks.FeeMarketQueryClient{}
+
+// Params
+func RegisterFeeMarketParams(feeMarketClient *mocks.FeeMarketQueryClient, height int64) {
+	feeMarketClient.On("Params", rpc.ContextWithHeight(height), &feemarkettypes.QueryParamsRequest{}).
+		Return(&feemarkettypes.QueryParamsResponse{Params: feemarkettypes.DefaultParams()}, nil)
+}
+
+func RegisterFeeMarketParamsError(feeMarketClient *mocks.FeeMarketQueryClient, height int64) {
+	feeMarketClient.On("Params", rpc.ContextWithHeight(height), &feemarkettypes.QueryParamsRequest{}).
+		Return(nil, sdkerrors.ErrInvalidRequest)
+}