+++ tests/integration_tests/test_types.py
@@ -1,11 +1,13 @@
 from web3 import Web3
 
 from .expected_constants import (
+    EXPECTED_ACCOUNT_PROOF,
     EXPECTED_FEE_HISTORY,
     EXPECTED_GET_PROOF,
     EXPECTED_GET_STORAGE_AT,
     EXPECTED_GET_TRANSACTION,
     EXPECTED_GET_TRANSACTION_RECEIPT,
+    EXPECTED_STORAGE_PROOF,
 )
 from .utils import (
     ADDRS,
@@ -47,8 +49,7 @@ def get_blocks(ethermint_rpc_ws, geth, with_transactions):
             with_transactions,
         ],
     )
-    res, err = same_types(eth_rsp, geth_rsp)
-    assert res, err
+    compare_types(eth_rsp, geth_rsp)
 
     # Get not existing block
     make_same_rpc_calls(
@@ -177,15 +178,18 @@ def test_get_storage_at(ethermint_rpc_ws, geth):
 
     contract = deploy_and_wait(w3)
     res = eth_rpc.make_request("eth_getStorageAt", [contract.address, "0x0", "latest"])
-    res, err = same_types(res["result"], EXPECTED_GET_STORAGE_AT)
-    assert res, err
+    compare_types(res["result"], EXPECTED_GET_STORAGE_AT)
 
 
-def send_and_get_hash(w3, tx_value=10):
+def send_tnx(w3, tx_value=10):
     # Do an ethereum transfer
     gas_price = w3.eth.gas_price
     tx = {"to": ADDRS["community"], "value": tx_value, "gasPrice": gas_price}
-    return send_transaction(w3, tx, KEYS["validator"])["transactionHash"].hex()
+    return send_transaction(w3, tx, KEYS["validator"])
+
+
+def send_and_get_hash(w3, tx_value=10):
+    return send_tnx(w3, tx_value)["transactionHash"].hex()
 
 
 def test_get_proof(ethermint_rpc_ws, geth):
@@ -195,27 +199,29 @@ def test_get_proof(ethermint_rpc_ws, geth):
     eth_rpc = w3.provider
     w3_wait_for_block(w3, 3)
     geth_rpc = geth.w3.provider
-    make_same_rpc_calls(
-        eth_rpc,
-        geth_rpc,
-        "eth_getProof",
-        ["0x57f96e6b86cdefdb3d412547816a82e3e0ebf9d2", ["0x0"], "latest"],
-    )
+    validator = ADDRS["validator"]
+    method = "eth_getProof"
+    for quantity in ["latest", "0x1024"]:
+        res = make_same_rpc_calls(
+            eth_rpc,
+            geth_rpc,
+            method,
+            [validator, ["0x0"], quantity],
+        )
+    res = send_tnx(w3)
 
-    make_same_rpc_calls(
-        eth_rpc,
-        geth_rpc,
-        "eth_getProof",
-        ["0x57f96e6b86cdefdb3d412547816a82e3e0ebf9d2", ["0x0"], "0x1024"],
-    )
+    proof = (eth_rpc.make_request(
+        method, [validator, ["0x0"], hex(res["blockNumber"])]
+    ))["result"]
+    compare_types(proof, EXPECTED_GET_PROOF["result"])
+    assert proof["accountProof"], EXPECTED_ACCOUNT_PROOF
+    assert proof["storageProof"][0]["proof"], EXPECTED_STORAGE_PROOF
 
     _ = send_and_get_hash(w3)
-
     proof = eth_rpc.make_request(
-        "eth_getProof", [ADDRS["validator"], ["0x0"], "latest"]
+        method, [validator, ["0x0"], "latest"]
     )
-    res, err = same_types(proof, EXPECTED_GET_PROOF)
-    assert res, err
+    compare_types(proof, EXPECTED_GET_PROOF)
 
 
 def test_get_code(ethermint_rpc_ws, geth):
@@ -233,8 +239,7 @@ def test_get_code(ethermint_rpc_ws, geth):
     contract = deploy_and_wait(w3)
     code = eth_rpc.make_request("eth_getCode", [contract.address, "latest"])
     expected = {"id": 4, "jsonrpc": "2.0", "result": "0x"}
-    res, err = same_types(code, expected)
-    assert res, err
+    compare_types(code, expected)
 
 
 def test_get_block_transaction_count(ethermint_rpc_ws, geth):
@@ -246,7 +251,7 @@ def test_get_block_transaction_count(ethermint_rpc_ws, geth):
     )
 
     make_same_rpc_calls(
-        eth_rpc, geth_rpc, "eth_getBlockTransactionCountByNumber", ["0x100"]
+        eth_rpc, geth_rpc, "eth_getBlockTransactionCountByNumber", ["0x1000"]
     )
 
     tx_hash = send_and_get_hash(w3)
@@ -259,8 +264,7 @@ def test_get_block_transaction_count(ethermint_rpc_ws, geth):
     )
 
     expected = {"id": 1, "jsonrpc": "2.0", "result": "0x1"}
-    res, err = same_types(block_res, expected)
-    assert res, err
+    compare_types(block_res, expected)
 
     make_same_rpc_calls(
         eth_rpc,
@@ -270,8 +274,7 @@ def test_get_block_transaction_count(ethermint_rpc_ws, geth):
     )
     block_res = eth_rpc.make_request("eth_getBlockTransactionCountByHash", [block_hash])
     expected = {"id": 1, "jsonrpc": "2.0", "result": "0x1"}
-    res, err = same_types(block_res, expected)
-    assert res, err
+    compare_types(block_res, expected)
 
 
 def test_get_transaction(ethermint_rpc_ws, geth):
@@ -289,8 +292,7 @@ def test_get_transaction(ethermint_rpc_ws, geth):
 
     tx_res = eth_rpc.make_request("eth_getTransactionByHash", [tx_hash])
 
-    res, err = same_types(EXPECTED_GET_TRANSACTION, tx_res)
-    assert res, err
+    compare_types(EXPECTED_GET_TRANSACTION, tx_res)
 
 
 def test_get_transaction_receipt(ethermint_rpc_ws, geth):
@@ -307,8 +309,7 @@ def test_get_transaction_receipt(ethermint_rpc_ws, geth):
     tx_hash = send_and_get_hash(w3)
 
     tx_res = eth_rpc.make_request("eth_getTransactionReceipt", [tx_hash])
-    res, err = same_types(tx_res, EXPECTED_GET_TRANSACTION_RECEIPT)
-    assert res, err
+    compare_types(tx_res, EXPECTED_GET_TRANSACTION_RECEIPT)
 
 
 def test_fee_history(ethermint_rpc_ws, geth):
@@ -322,8 +323,7 @@ def test_fee_history(ethermint_rpc_ws, geth):
     _ = send_and_get_hash(w3)
     fee_history = eth_rpc.make_request("eth_feeHistory", [4, "latest", [100]])
 
-    res, err = same_types(fee_history, EXPECTED_FEE_HISTORY)
-    assert res, err
+    compare_types(fee_history, EXPECTED_FEE_HISTORY)
 
 
 def test_estimate_gas(ethermint_rpc_ws, geth):
@@ -338,11 +338,19 @@ def test_estimate_gas(ethermint_rpc_ws, geth):
     make_same_rpc_calls(eth_rpc, geth_rpc, "eth_estimateGas", [{}])
 
 
+def compare_types(actual, expected):
+    res, err = same_types(actual, expected)
+    if not res:
+        print(err)
+        print(actual)
+        print(expected)
+    assert res, err
+
+
 def make_same_rpc_calls(rpc1, rpc2, method, params):
     res1 = rpc1.make_request(method, params)
     res2 = rpc2.make_request(method, params)
-    res, err = same_types(res1, res2)
-    assert res, err
+    compare_types(res1, res2)
 
 
 def test_incomplete_send_transaction(ethermint_rpc_ws, geth):