+++ tests/integration_tests/utils.py
@@ -80,7 +80,7 @@ def wait_for_new_blocks(cli, n):
 
 
 def wait_for_block(cli, height, timeout=240):
-    for i in range(timeout * 2):
+    for _ in range(timeout * 2):
         try:
             status = cli.status()
         except AssertionError as e:
@@ -96,7 +96,7 @@ def wait_for_block(cli, height, timeout=240):
 
 
 def w3_wait_for_block(w3, height, timeout=240):
-    for i in range(timeout * 2):
+    for _ in range(timeout * 2):
         try:
             current_height = w3.eth.block_number
         except Exception as e:
@@ -127,7 +127,7 @@ def deploy_contract(w3, jsonfile, args=(), key=KEYS["validator"]):
     acct = Account.from_key(key)
     info = json.loads(jsonfile.read_text())
     contract = w3.eth.contract(abi=info["abi"], bytecode=info["bytecode"])
-    tx = contract.constructor(*args).buildTransaction({"from": acct.address})
+    tx = contract.constructor(*args).build_transaction({"from": acct.address})
     txreceipt = send_transaction(w3, tx, key)
     assert txreceipt.status == 1
     address = txreceipt.contractAddress