+++ types/validation.go
@@ -3,7 +3,8 @@ package types
 import (
 	"bytes"
 
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	errorsmod "cosmossdk.io/errors"
+	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/ethereum/go-ethereum/common"
 )
 
@@ -20,8 +21,8 @@ func IsZeroAddress(address string) bool {
 // ValidateAddress returns an error if the provided string is either not a hex formatted string address
 func ValidateAddress(address string) error {
 	if !common.IsHexAddress(address) {
-		return sdkerrors.Wrapf(
-			sdkerrors.ErrInvalidAddress, "address '%s' is not a valid ethereum hex address",
+		return errorsmod.Wrapf(
+			errortypes.ErrInvalidAddress, "address '%s' is not a valid ethereum hex address",
 			address,
 		)
 	}
@@ -32,8 +33,8 @@ func ValidateAddress(address string) error {
 // formatted string address or is equal to zero
 func ValidateNonZeroAddress(address string) error {
 	if IsZeroAddress(address) {
-		return sdkerrors.Wrapf(
-			sdkerrors.ErrInvalidAddress, "address '%s' must not be zero",
+		return errorsmod.Wrapf(
+			errortypes.ErrInvalidAddress, "address '%s' must not be zero",
 			address,
 		)
 	}